Resolves https://github.com/Rythe-Interactive/Legion-LLRI/pull/56#discussion_r728926382

+++ legion/engine/llri-vk/utils.cpp
@@ -143,7 +143,7 @@ namespace LLRI_NAMESPACE
                     return static_cast<uint32_t>(i);
             }
 
-            return -1;
+            return static_cast<uint32_t>(-1);
         }
 
         /**
